Add a simple version of the whole mocker flow
At the moment the flow only generates one file with only one table. We
of course want to support multiple tables, but as of now this isn't
possible. Either the generators need to be rewritten to take multiple
tables into account (at the moment it is one table per generator) or the
creating of the files and calling the generators needs to be rewritten,
changing a single output file to a output folder with then includes
multiple files in the format of `{table_name}.{type}`. If the last
option is chosen, the generators don't need to be changed.